web-app/requirements.txt
Flask==2.2.2
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
pynytimes==0.4.0
python-dotenv==0.21.0
requests==2.28.1